+++ src/components/AccountBox.jsx
@@ -4,7 +4,7 @@ const AccountBox = () =>{
     return(
         <div className="account-balance-box">
            <p>Balance actual (Bs.)</p>
-            <p>16,57 Total Bs.</p>
+            <p>16,57 Bs.</p>
         </div>
     );
 }
