--- src/components/Seccion start/Start.css
@@ -1,91 +1,0 @@
-.section-start{
-    display: flex;
-    flex-direction: row;
-    margin: 0 auto;
-    height: auto;
-    align-items: center; 
-}
-
-.text-section{
-    width: 45%;
-    text-align: center;
-    margin-top: 8%;
-    color:#085f63 ;
-    display: flex;
-    flex-direction: column;
-    justify-content: center;
-
-}
-
-.text-section h1{
-    line-height: 1.2; 
-    font-size: 2.5rem; 
-    font-weight: bold; 
-}
-
-.text-section h2{
-    font-size: 2rem; 
-    margin: 10px 0; 
-}
-
-hr{
-    width: 60%;
-    border: 3px solid #000000;
-    margin: 5% auto;
-}
-
-.img-section{
-    width: 55%;
-    display: flex;
-    justify-content: center;
-}
-
-.img-lady{
-    max-width: 100%;
-    height: auto;
-}
-@media (max-width: 853px) {
-    .text-section,
-    .img-section {
-        width: 100%; 
-    }
-    .img-lady {
-        width: 100%; 
-        margin-top: 20px; 
-    }
-
-  .text-section h1{
-    font-size: 1.8rem;
-  }
-  .text-section h2 {
-    font-size: 1.3rem;
-    margin: 5% 0;
-}
-  hr{
-    width: 80%;
-  }
-}
-
-@media (max-width: 380px) {
-    .text-section,
-    .img-section {
-        width: 100%; 
-    }
-
-    .img-lady{
-    width: 100%;
-    }
-
-    .text-section h1 {
-        font-size: 1rem; 
-        line-height: 1.2;
-    }
-    .text-section h2 {
-        font-size: 0.8rem; 
-    }
-    hr {
-        width: 80%; 
-    }
-
-}
-
