--- src/components/Seccion start/Start.jsx
@@ -1,10 +1,0 @@
-import "./Start.css";
-
-
-const Start = () =>{
-    return(
-      
-    );
-}
-
-export default Start;
