+++ src/components/styles/NavBarBanking.css
@@ -1,7 +1,6 @@
 .navbar-banking {
     align-items: center;
     background-color: #072146;
-    
     color: #fff;
   }
 
@@ -18,6 +17,7 @@
 
   .navbar-links{
     display: flex;
+    background-color: rgba(255, 255, 255, 0.809);
     justify-content: space-between;
     box-sizing: border-box;
     padding: 0.1rem 5rem;
